requirements.txt
streamlit
langchain
langgraph
langchain-openai
langchain-community
python-dotenv
pandas
mcp
pydantic
